Add a new VSPackage for the Project System.
While a Dylan application template is now visible in the New Project
dialog the project file (.dylproj) it generates does not contain the
data the project system expects it to have. A bit of work is still
required in that area before a new Dylan application project can
actually be successfully created.