Add basic General project property page.
This is one of the pages that appear when a user right-clicks on a Dylan
project in the Solution Explorer and selects Properties from the
context-menu. Currently it only contains a few fields that will be
mapped to those typically found in Dylan LID files. The current
grid-based implementation might get replaced by a user control that will
provide a look that's more consistent with VS2012 property pages.